+++ core-ng/src/main/java/core/framework/db/Repository.java
@@ -16,7 +16,7 @@ public interface Repository<T> {
         return query.fetch();
     }
 
-    default int count(String where, Object... params) {
+    default long count(String where, Object... params) {
         Query<T> query = select();
         if (where != null) query.where(where, params);
         return query.count();
